Add files via upload
This is just a simple example of how to create a password locker using a dictionary in Python.

This program first creates a dictionary to store the passwords. Then, it defines three functions: generate_password() to generate a random password, add_password() to add a password to the locker, and get_password() to get a password from the locker. The delete_password() function is also defined, but it is not used in this program.

The main function of the program first generates a random password and then adds it to the locker. It then gets the password for the "email" account and prints it out. Finally, it deletes the password for the "bank" account and prints the contents of the locker.